viam: Add GraphVisitor interface in vadl.viam.graph
This commit includes a new GraphVisitor interface in the vadl.viam.graph package, which is designed to visit nodes in a graph and obtain a result of a specific type. An Applier interface is also added inside GraphVisitor, assigned with the task of applying new input values to specific nodes in the graph.