viam: Add GraphVisitor logic to Node
Introduced GraphVisitor and its logic to the Node class and its subclasses. The GraphVisitor traverses the graph visiting each node and applies specified logic on every input. The subclasses, TestNodeWithInput and TestNodeWithTwoInputs, have been updated to override the applyOnInputs() method of the Node and incorporate this logic to their own inputs.