viam: Update type cast elimination to fix constant evaluation
The previous rule that if the target type is a signed integer, the cast should be a sign extension leads to wrong constant evaluation, as the literal 2 has the type UInt<2> which is 0b10.
However, when casting this to e.g. SInt<4>, this causes a sign extension and therefore a wrong number.