iss: Add the IssTemplateRenderingPass and IssDefaultRenderingPass
The IssTemplateRenderingPass extends the AbstractTempalateRenderingPass and sets some default variables.
It also automatically determines the output path, based on the template path, by replacing all `gen-arch` substrings by the concrete architecture names from the configuration.
The IssDefaultRenderingPass takes a issTemplatePath and fills the template with the default variables set by the AbstractTemplateRenderingPass.