viam: Improve HTML dump
Unfortunately the graphviz library we are using only allows pixel values
for the size. However we want the svg to fill it's container which
only ever kinda worked in Chrome (but threw errors) and just didn't
work in Firefox.

Now we don't set the svg size via the graphviz library but with custom
CSS which works in all browsers without any errors.

Graphviz library documentation to height and width:
https://github.com/magjac/d3-graphviz?tab=readme-ov-file#graphviz_width