parser: Fix multiple invalid locations
There were multiple issues with how locations (especially the start)
were generated.

One was that `loc` was called before the first token was consumed which
lead to the location spanning to the last token of the previous rule.
(This was the case for instructions and encodings).

Another one was that multiple tokens were consumed before `loc` was
called leading to a location that didn't include the whole rule.

Now, before we consume a single token `lookaheadLoc` is called which
will always include the first token of the rule.
This easily avoid mistakes from copy and paste and is more resilient to
changes in the grammar.