parser: Refactor Ungrouper to modify nodes in-place
Recreating nodes in the Ungrouper is not only slow, more memory
intensive but also incorrect.

The issue is that the symbol resolver runs before the Ungrouper.
Which means that if the Ungrouper creates new nodes, the symbol table
will point to an outdated node that is no longer part of the AST.

This becomes especially problematic when the the typechecker adds types
to nodes in the AST but the symboltable points to nodes outside the
AST that haven't been checked.

Moreover, there is simply no reason to recreate/replace nodes.
I assume this was mostly done because records are easy to use, but are
immutable and therefore cannot be modified in-place.