typechecker: Implement new semantics for constant branches
Branching expression where are all branches return constants are hard to
type. Instead of throwing a generic error, for constant definitions we
are now evaluating the condition to know which branch is selected and
propagating the type of that branch.

When the condition cannot be evaluated we display a more tailored
diagnostic with the hint that in the future this will be solved.