iss-testing: Refactor ISS instruction test framework
This refactoring builds the test framework from ground up, with a much simpler and more customizable approach.
Instead of using QMP to communicate with QEMU to get register information and shutdown, we use HTIF to shut down an ISS test and use a custom plugin to read registers right before shutdown.