lcb: Handle signed and unsigned const mat #129
The header class was copied from RISCV. The difference is
that upstream is only storing machine instructions which
have smaller bit widths in the instructions. We store
pseudo instructions which hold the whole immediate. Therefore,
the type change in the header.

Everything is a signed integer. Even though, it is not. So if
no signed constant sequences fits then try the unsigned constant
sequences.