iss-testing: Refactor test framework (#140)
* iss-testing: Refactor ISS instruction test framework

This refactoring builds the test framework from ground up, with a much simpler and more customizable approach.
Instead of using QMP to communicate with QEMU to get register information and shutdown, we use HTIF to shut down an ISS test and use a custom plugin to read registers right before shutdown.

* iss-testing: Add endstate.c plugin

This prints the final register state of the simulated CPU, which is used for comparison.

* iss-testing: Use new test framework with QEMU plugin

This integrates the new framework and the QEMU plugin to the java testing side.

* iss-testing: Move risc-v class in own package