typechecker: Fix crash on unfound symbol
The problem was again that the symbol resolver missed some cases.
I'm more and more convinced that the manual iterating through the AST
in each visitor is error prone and for passes that don't care to much
about the order we should provide one default implementation.

But that's work for another commit, for now this commit fixes the
missing children in the symbol resolver.