lcb: Refactor LinkerComponents and emit Fixups
Refactor the linker components to no longer follow the strict one-to-one relationship of Modifier-VariantKind-Fixup-ElfRelocation for each relocation. Now there is a one-to-one relation of Modifier-VariantKind for user defined- and automatically generated relocations. Then for each of these there are multiple Fixups and for each Fixup then there is exactly one ElfRelocation.