gcb+lcb: Refactor linker components and emit relocations (#177)
* lcb: Generate one VariantKind per user defined relocation

* lcb: Refactor LinkerComponents and emit Fixups

Refactor the linker components to no longer follow the strict one-to-one relationship of Modifier-VariantKind-Fixup-ElfRelocation for each relocation. Now there is a one-to-one relation of Modifier-VariantKind for user defined- and automatically generated relocations. Then for each of these there are multiple Fixups and for each Fixup then there is exactly one ElfRelocation.

* lcb: Adjust tests to new variant kinds and relocations

* asm: Check emitted fixups in CodeEmitter tests

- Add LA to rv32im asm grammar
- Fix InstExpander test after rebase

* lcb: Add test for GenerateLinkerComponentsPass

Remove dead code in EmitISelLoweringPass