parser: Improve location utility function naming
Many people that started contributing to the parser got tricked by the
confusingly named function `loc()` which only returned the location of
the last consumed token instead of the more intuitive guess of the
location of the whole rule.

The better naming should now make it much clearer what the functions do.