frontend: Rename ExtendId and IdToStr to AsId and AsStr
In the very first implementation of VADL's macro system this function took one String argument and extended an identifier with this string. OpenVADL's version takes any number of identifiers and strings and creates an identifier with the name resulting from converting all arguments to strings, concatenating them and converting the concatenated string to an identifier.
gh: #103