frontend: Fix assembly with multiple instructions
Assembly definitions can have multiple (pseudo) instructions for which
they are implemented. The expression in the assembly can reference
values/variables from the instructions which might bind to different
definitions and types in the instructions.

So to better handle those cases such assembly definitions are now
expanded to multiple definitions (one for each instruction that is
implemented).

This means that assembly definitions behave more like macros which is
the reason why most of this expanding is now implemented in the
Macroexpander.

gh: #304