lcb: Refactored encoding/predicate/access function generation
Before this commit, we had multiple passes which essentially
did the same. The idea is to have a single pass for
the immediates and all the others take that as an input.
Additionally, function generation should be already in that pass.
So everything related to immediates is in one pass.