lcb: Use correct source for field access functions
In `LlvmPseudoInstructionLoweringUnconditionalJumpsStrategyImpl` we have
a conditional jump and try to make it unconditional. We check
whether the field access function gets assigned. The bug was
that the check counted the number of field access functions in the
format. But it does not matter how many field access functions are
used. Only the used field access functions matter.