parser: Adapt new forall expr, statement syntax
The iterating variable can now have an optional cast, which for the
initial implementation in the typechecker won't be _that_ optional
since we can only infer it once bidirectional type checking is
implemented.

Also the `append` variant for the forall expression is diabled for
now as we aren't sure that it will be used since we are still unsure
how protocols are implemented.