rtl: Synthesize control and forward logic
Implement control/forward logic and connect it to the stages
by modifying read/write behavior.
Emit Chisel code from the MiA and its components in the form of
Chisel modules that instantiate and connect the stages and logic
elements inside a top module. All external resource (memory)
are provided as IO ports on the top module.
This also generates a Chisel test that instantiates the core
and connects the IO ports to a simulated memory. The test also
can load ELF files.