rtl: Improve MiA synthesis determinism
The use of HashMaps and HashSets in a lot of steps of the MiA synthesis
and the HDL generation caused node IDs and probably also synthesis
results to differ between executions. This is solved by using
LinkedHashMaps and LinkedHashSets in places where we iterate over these
maps and sets. Also improves the naming of registers, signals and ports.